Revert "Simplify inception provider module to use openai streaming api"
This reverts commit 01c93176cfb6ad263f7022c3da8a0ffedeb12427.
