+++ rig-core/examples/agent_with_inception.rs
@@ -2,18 +2,21 @@ use std::env;
 
 use rig::{
     completion::Prompt,
-    providers::inception::{Client, MERCURY_CODER_SMALL},
+    providers::inception::{ClientBuilder, MERCURY_CODER_SMALL},
 };
 
 #[tokio::main]
 async fn main() -> Result<(), anyhow::Error> {
     // Create Inception Labs client
-    let client = Client::new(&env::var("INCEPTION_API_KEY").expect("INCEPTION_API_KEY not set"));
+    let client =
+        ClientBuilder::new(&env::var("INCEPTION_API_KEY").expect("INCEPTION_API_KEY not set"))
+            .build();
 
     // Create agent with a single context prompt
     let agent = client
         .agent(MERCURY_CODER_SMALL)
         .preamble("You are a helpful AI assistant.")
+        .temperature(0.0)
         .build();
 
     // Prompt the agent and print the response
