+++ rig-core/examples/inception_streaming.rs
@@ -1,5 +1,5 @@
 use rig::{
-    providers::inception::{self, MERCURY_CODER_SMALL},
+    providers::inception::{self, completion::MERCURY_CODER_SMALL},
     streaming::{stream_to_stdout, StreamingPrompt},
 };
