+++ rig-core/src/providers/inception/mod.rs
@@ -1,0 +1,6 @@
+pub mod client;
+pub mod completion;
+pub mod streaming;
+
+pub use client::{Client, ClientBuilder};
+pub use completion::MERCURY_CODER_SMALL;
